Assignment 0 added
This is the "index.html" file the assignment asks for. I've been putting together my website with an index and navigation and the like, so I've got an index for my assignments and this is assignment 0. Follow the link to my website in the readme to check it out. 
Sorry if this is inconvenient for grading. I'll discuss it with you in lecture on Monday.